Remoção de bugs e melhoria de layout

+++ src/app/segments/list-segments/list-segments.component.ts
@@ -4,14 +4,18 @@ import { Table } from 'primeng/table';
 import { FormBuilder, FormGroup, Validators, FormControl } from '@angular/forms';
 import { SegmentService } from 'src/app/share/services/segment/segment.service';
 import { Subscription } from 'rxjs';
+import { MessageService } from 'primeng/api';
+import { ConfirmationService, Message } from 'primeng/api';
+
 
 
 @Component({
   selector: 'app-list-segments',
   templateUrl: './list-segments.component.html',
-  styleUrls: ['./list-segments.component.css']
+  styleUrls: ['./list-segments.component.css'],
+  providers: [ConfirmationService, MessageService]
 })
-export class ListSegmentsComponent implements OnInit,OnDestroy {
+export class ListSegmentsComponent implements OnInit, OnDestroy {
 
   public segments: Segment[] = [];
   displayModal: boolean = false;
@@ -21,19 +25,22 @@ export class ListSegmentsComponent implements OnInit,OnDestroy {
   headerDialog: string = 'Novo Segmento';
   subseg!: Subscription;
 
+
   constructor(private segmentService: SegmentService,
-    private formBuilder: FormBuilder) {
+    private formBuilder: FormBuilder,
+    private confirmationService: ConfirmationService,
+    private messageService: MessageService) {
     this.segmentForm = this.formBuilder.group({
       name: [null, [Validators.required, Validators.minLength(5), Validators.maxLength(50)]],
       tax: [null, [Validators.required, Validators.pattern(this.numRegex)]],
       isActive: [null],
-    });    
+    });
   }
   ngOnDestroy(): void {
     this.subseg.unsubscribe();
   }
 
-  novoSegmento(){
+  novoSegmento() {
     this.segmentSelect = {} as Segment;
     this.segmentForm.reset();
     this.headerDialog = 'Novo Segmento';
@@ -52,21 +59,24 @@ export class ListSegmentsComponent implements OnInit,OnDestroy {
 
   loadSegments() {
     this.segmentService.get();
-    this.subseg = this.segmentService.segmentsObs.subscribe(segs =>{
+    this.subseg = this.segmentService.segmentsObs.subscribe(segs => {
       this.segments = segs
     });
   }
   onSubmit() {
     if (this.segmentForm.valid) {
       if (this.segmentSelect.id) {
-        const seg: SegmentPut = {         
+        const seg: SegmentPut = {
           name: this.segmentForm.value.name,
           tax: parseFloat(this.segmentForm.value.tax),
           isActive: this.segmentForm.value.isActive == 1
         };
-        this.segmentService.put(this.segmentSelect.id,seg).then(segs => {
-         this.loadSegments();
-         this.segmentSelect = {} as Segment;
+        this.segmentService.put(this.segmentSelect.id, seg).then(segs => {
+          this.loadSegments();
+          this.messageService.add({ severity: 'success', summary: 'Alteração', detail: 'Segmento alterado com sucesso' });
+          this.segmentSelect = {} as Segment;
+        }).catch(error => {
+          this.messageService.add({ severity: 'error', summary: 'Erro na alteração', detail: 'Consulte adm do sistema' });
         });
       }
       else {
@@ -75,8 +85,14 @@ export class ListSegmentsComponent implements OnInit,OnDestroy {
           tax: parseFloat(this.segmentForm.value.tax.replace(',', '.'))
         };
 
-        this.segmentService.post(seg).then(segs =>
-          this.loadSegments());
+        this.segmentService.post(seg).then(segs => {
+          this.loadSegments();
+          this.messageService.add({ severity: 'success', summary: 'Inserção', detail: 'Segmento inserido com sucesso' });
+        }
+        ).catch(error => {
+          this.messageService.add({ severity: 'error', summary: 'Erro na inclusão', detail: 'Consulte adm do sistema' });
+        });
+
       }
       this.displayModal = false;
       this.segmentForm.reset();
@@ -89,8 +105,21 @@ export class ListSegmentsComponent implements OnInit,OnDestroy {
   }
 
   onRowRemove(event: any) {
-   this.segmentService.delete(event.id).then(segs =>
-    this.loadSegments());
+    this.confirmationService.confirm({
+      message: 'Deseja realmente excluir esse Segmento?',
+      header: 'Confirmação',
+      icon: 'pi pi-info-circle',
+      accept: () => {
+        this.segmentService.delete(event.id).then(segs => {
+          this.loadSegments();
+          this.messageService.add({ severity: 'success', summary: 'Confirmado', detail: 'Segmento excluido' });
+        }
+        ).catch(error => {
+          this.messageService.add({ severity: 'error', summary: 'Erro na exclusão', detail: 'Consulte adm do sistema' });
+        });
+      }
+    });
+
   }
 
 }
